Add start_tls hostname param (alias: sni_hostname)
`hostname` is what OpenSSL::SSL::SSLConnection calls it, and I'd prefer
to use the same names as stdlib whenever possible.

This adds documentation to #start_tls for both options.  Previously
there was no documentation for :sni_hostname, so I've documented
`hostname` as the preferred name and `sni_hostname` as the alias for
backward compatibility.